test(gitops): address review feedback on repairBrokenSubmodules tests (#608)
- dropped the extra git add/commit in 'broken submodule repaired', use git ls-files --stage to check the index directly after repair
- strengthen 'broken gitlink file removed' to assert 160000 in the index rather than inspecting the stale .git file content
- move git test helpers into testutil_test.go so they can be shared by upcoming tests for findDirtyRepos and convertToSubmodules
